Changes over parameter parsing
